refactor(editor): Remove `i18n` and `toast` usage from setting store (no-changelog) (#16721)
